- Created getUserFavorites in FavoriteController - Deleted getUserFavorites from UserController - Deleted .gitkeep
